Allow workflow to request resources via hpc-resource-provisoner's api
Before a worker is launched, the workflow requests a cluster formation,
and if not already available it waits for the resources to be provisioned.

When a resource is available, the get request returns the ssh key and head
node ip to allow connecting to it and launch jobs.

The workflow does not release the resource as we want it to persist between
workflow runs and be reused.

Some notes:

* the provisioner api does not yet allow to specify a profile
  (configuration template of cloud formation)

* the provisioner api relies on (project, vlab_id) idempotence and to my
  knowledge doesn't also depend on the profile of the cloud formation